refactor(worker): flatten crate structure and remove legacy V1 activities
- Move v2/agent_workflow.rs to root as agent_workflow.rs
- Merge v2/activities.rs into root activities.rs with V2 activities only
- Remove legacy V1 activity types from types.rs (LoadAgentInput, CallLlmInput, etc.)
- Remove legacy V1 activity routing from worker.rs
- Remove v2/ directory entirely
- Update all imports and re-exports

The worker crate now has a flat structure with:
- agent_workflow.rs: AgentWorkflow state machine
- activities.rs: load_agent, call_model, execute_tool activities using Atoms
- types.rs: WorkflowAction, workflow_names, constants
- worker.rs: Temporal polling and activity execution

This completes the migration from session_workflow to agent_workflow.
