+++ crates/everruns-worker/src/lib.rs
@@ -1,18 +1,18 @@
 pub mod activities;
 pub mod adapters;
+pub mod agent_workflow;
 pub mod client;
 pub mod runner;
 pub mod types;
 pub mod unified_tool_executor;
-pub mod v2;
 pub mod worker;
 pub mod workflow_registry;
 pub mod workflow_traits;
 
 // Re-export main types
+pub use agent_workflow::{AgentWorkflow, AgentWorkflowInput};
 pub use runner::{create_runner, run_worker, AgentRunner, RunnerConfig, TemporalRunner};
 pub use types::WorkflowAction;
-pub use v2::{AgentWorkflow, AgentWorkflowInput};
 pub use worker::TemporalWorker;
 pub use workflow_registry::{WorkflowFactory, WorkflowRegistry, WorkflowRegistryBuilder};
 pub use workflow_traits::{Workflow, WorkflowInput};
